support writing to deltalake using ray and remove buffer.arrow in FAST stream_mode

+++ pfeed/_io/buffer_io.py
@@ -16,8 +16,6 @@ from pfeed.enums import StreamMode
 
 class BufferIO(BaseIO):
     BUFFER_FILENAME = 'buffer.arrow'
-    # REVIEW: make it a param?
-    MAX_BUFFER_SIZE = 100  # number of data
     
     def __init__(
         self,
@@ -75,37 +73,54 @@ class BufferIO(BaseIO):
             del self._buffer_files[file_path]
     
     def write(self, file_path: Path, streaming_data: dict, schema: pa.Schema):
-        self._buffer.append(streaming_data)
         if self._stream_mode == StreamMode.FAST:
-            if len(self._buffer) >= self.MAX_BUFFER_SIZE:
-                self.spill_to_disk(file_path, schema)
+            self._buffer.append(streaming_data)
         elif self._stream_mode == StreamMode.SAFE:
-            self.spill_to_disk(file_path, schema)
+            self._spill_to_disk(file_path, streaming_data, schema)
         else:
             raise ValueError(f'Invalid stream mode: {self._stream_mode}')
             
-    # OPTIMIZE
-    def spill_to_disk(self, file_path: Path, schema: pa.Schema):
+    def _spill_to_disk(self, file_path: Path, streaming_data: dict, schema: pa.Schema):
         '''Spill in-memory buffer to buffer.arrow'''
         self._open_writers(file_path, schema)
-        row = { name: [ streaming_data[name] for streaming_data in self._buffer ] for name in schema.names }  # make each column a list
+        row = { name: [ streaming_data[name] ] for name in schema.names }  # make each column a list
         batch = pa.record_batch(row, schema=schema)
         self._ipc_writers[file_path].write_batch(batch)
         buffer_file = self._buffer_files[file_path]
         buffer_file.flush()  # flush python buffer to OS buffer
         os.fsync(buffer_file.fileno())  # flush OS buffer to disk
         # write_size = os.fstat(buffer_file.fileno()).st_size
-        self._buffer.clear()
             
-    def read(self, file_path: Path) -> pa.Table:
-        self._close_writers(file_path)
-        with open(file_path, "rb") as f:
-            with pa.ipc.open_stream(f) as reader:
-                table = reader.read_all()
-                table = table.combine_chunks()
-                return table
+    def read(self, file_path: Path, schema: pa.Schema | None=None) -> pa.Table:
+        if self._stream_mode == StreamMode.FAST:
+            assert schema is not None, 'schema is required for reading in fast mode'
+            return pa.Table.from_pylist(self._buffer, schema=schema)
+        elif self._stream_mode == StreamMode.SAFE:
+            self._close_writers(file_path)
+            with open(file_path, "rb") as f:
+                with pa.ipc.open_stream(f) as reader:
+                    table = reader.read_all()
+                    # REVIEW: combine_chunks() could create misaligned memory for Delta Lake FFI
+                    # currently if only writing one data per batch, this issue doesn't occur for some reason
+                    table = table.combine_chunks()
+                    '''More details on the problem:
+                    Because the combined table's memory layout doesn't meet the 8-byte alignment requirements for Delta Lake's Rust FFI interface
+                    The problem occurs when:
+                    1. Multiple small record batches get combined
+                    2. combine_chunks() creates a memory layout where buffer addresses aren't 8-byte aligned
+                    3. Delta Lake's Rust FFI expects all Arrow buffer pointers to be aligned to 8-byte boundaries
+                    Instead, reconstruct the table to ensure proper memory alignment
+                    # NOTE: fallback solution when this issue occurs again
+                    if table.num_rows > 0:
+                        df = table.to_pandas()
+                        table = pa.Table.from_pandas(df, preserve_index=False, schema=table.schema)
+                    '''
+                    return table
 
-    def clear_disk(self, file_path: Path):
-        '''Clear the buffer.arrow file'''
-        if file_path.exists():
-            file_path.unlink()
+    def clear(self, file_path: Path):
+        '''Clear buffer or the buffer.arrow file'''
+        if self._stream_mode == StreamMode.FAST:
+            self._buffer.clear()
+        elif self._stream_mode == StreamMode.SAFE:
+            if file_path.exists():
+                file_path.unlink()
